requirements.txt
absl-py==2.2.2
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiohttp-retry==2.9.1
aiosignal==1.4.0
arabic-reshaper==3.0.0
asgiref==3.9.2
astunparse==1.6.3
attrs==25.4.0
blinker==1.9.0
boto3==1.38.7
botocore==1.38.7
certifi==2025.4.26
chardet==5.2.0
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
distlib==0.4.0
Django==5.2.6
djangorestframework==3.16.1
et_xmlfile==2.0.0
filelock==3.18.0
filetype==1.2.0
fire==0.7.0
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
flatbuffers==25.2.10
fonttools==4.56.0
frozenlist==1.8.0
fsspec==2025.3.2
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google-pasta==0.2.0
greenlet==3.1.1
grpcio==1.71.0
h5py==3.13.0
huggingface-hub==0.24.7
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
keras==3.9.2
kiwisolver==1.4.8
libclang==18.1.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
ml_dtypes==0.5.1
mpmath==1.3.0
multidict==6.7.0
namex==0.0.9
networkx==3.4.2
numpy==2.2.2
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.15.0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pillow_heif==0.22.0
platformdirs==4.4.0
propcache==0.4.1
protobuf==5.29.4
psutil==7.0.0
py-cpuinfo==9.0.0
pybboxes==0.1.6
pygame==2.6.1
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.2.1
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.1
PyYAML==6.0.2
reportlab==4.3.1
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
roboflow==1.1.63
runserver==1.0.4
s3transfer==0.12.0
sahi==0.11.22
scipy==1.15.2
seaborn==0.13.2
setuptools==80.1.0
shapely==2.1.0
six==1.17.0
smmap==5.0.2
SQLAlchemy==2.0.38
sqlparse==0.5.3
sympy==1.14.0
telepot==12.7
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow_cpu==2.19.0
termcolor==3.1.0
terminaltables==3.1.10
thop==0.1.1.post2209072238
torch==2.7.0
torchvision==0.22.0
tqdm==4.67.1
twilio==9.8.6
typing_extensions==4.12.2
tzdata==2025.1
ultralytics==8.3.124
ultralytics-thop==2.0.14
urllib3==2.4.0
virtualenv==20.34.0
Werkzeug==3.1.3
wheel==0.45.1
windows-curses==2.4.1
wrapt==1.17.2
XlsxWriter==3.2.2
yarl==1.22.0
yolov5==7.0.14
